requirements.txt
django
gunicorn
django-heroku
requests
djangorestframework
markdown
django-filter
Pillow